Add files via upload
For first while we print n output will the whole n matrix. To print the elements in row wise we separately print each nested list by using for loop ( if r= 0 it will print first nested loop , when r=1 it will print next and so on ). To print the elements in matrix wise we have to make each elements of nested loop separately and print each one by one ( assign the first nested loop as i=0 ([10 , 20 , 30 ]) and the first element of the loop   j=0 ( 10 ) . By using nested for loop first check the range of i ( from 0 to 2 ) and range of j ( from 0 to length of i ) and print each elements of nested list separately using end = ' ' function